Move Contents classes into separate translation units
