+++ entity/contents/BUILD.gn
@@ -1,0 +1,30 @@
+# Copyright 2013 The Flutter Authors. All rights reserved.
+# Use of this source code is governed by a BSD-style license that can be
+# found in the LICENSE file.
+
+import("//flutter/impeller/tools/impeller.gni")
+
+impeller_component("contents") {
+  sources = [
+    "clip_contents.cc",
+    "clip_contents.h",
+    "contents.cc",
+    "contents.h",
+    "linear_gradient_contents.cc",
+    "linear_gradient_contents.h",
+    "solic_color_contents.h",
+    "solid_color_contents.cc",
+    "text_contents.cc",
+    "text_contents.h",
+    "texture_contents.cc",
+    "texture_contents.h",
+  ]
+
+  public_deps = [
+    "../../archivist",
+    "../../image",
+    "../../renderer",
+    "../../typographer",
+    "../:entity_shaders",
+  ]
+}
