+++ entity/contents/contents.h
@@ -1,0 +1,37 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#pragma once
+
+#include <functional>
+#include <memory>
+#include <vector>
+
+#include "flutter/fml/macros.h"
+
+namespace impeller {
+
+class ContentContext;
+class ContentContextOptions;
+class Entity;
+class Surface;
+class RenderPass;
+
+ContentContextOptions OptionsFromPass(const RenderPass& pass);
+
+class Contents {
+ public:
+  Contents();
+
+  virtual ~Contents();
+
+  virtual bool Render(const ContentContext& renderer,
+                      const Entity& entity,
+                      RenderPass& pass) const = 0;
+
+ private:
+  FML_DISALLOW_COPY_AND_ASSIGN(Contents);
+};
+
+}  // namespace impeller
