+++ entity/contents/linear_gradient_contents.h
@@ -1,0 +1,49 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#pragma once
+
+#include <functional>
+#include <memory>
+#include <vector>
+
+#include "flutter/fml/macros.h"
+#include "impeller/entity/contents/contents.h"
+#include "impeller/entity/solid_stroke.vert.h"
+#include "impeller/geometry/color.h"
+#include "impeller/geometry/path_component.h"
+#include "impeller/geometry/point.h"
+#include "impeller/geometry/rect.h"
+#include "impeller/renderer/texture.h"
+#include "impeller/typographer/glyph_atlas.h"
+#include "impeller/typographer/text_frame.h"
+
+namespace impeller {
+
+class LinearGradientContents final : public Contents {
+ public:
+  LinearGradientContents();
+
+  ~LinearGradientContents() override;
+
+  // |Contents|
+  bool Render(const ContentContext& renderer,
+              const Entity& entity,
+              RenderPass& pass) const override;
+
+  void SetEndPoints(Point start_point, Point end_point);
+
+  void SetColors(std::vector<Color> colors);
+
+  const std::vector<Color>& GetColors() const;
+
+ private:
+  Point start_point_;
+  Point end_point_;
+  std::vector<Color> colors_;
+
+  FML_DISALLOW_COPY_AND_ASSIGN(LinearGradientContents);
+};
+
+}  // namespace impeller
