+++ entity/entity_pass_delegate.h
@@ -7,7 +7,7 @@
 #include <memory>
 
 #include "flutter/fml/macros.h"
-#include "impeller/entity/contents.h"
+#include "impeller/entity/contents/contents.h"
 #include "impeller/renderer/texture.h"
 
 namespace impeller {
